yoloface/requirements2.txt
scipy==1.7.2
seaborn==0.11.1
joblib==1.0.1
matplotlib==3.4.3
pandas==1.3.4
opencv_python==4.5.3.56
tqdm==4.62.3
numpy==1.20.3
requests==2.25.1
onnx==1.10.2
Pillow==9.0.0
PyYAML==6.0
wandb==0.12.9
